add map method to cards listing
