+++ src/Homes/All/Filters/Guests.js
@@ -1,8 +1,8 @@
 import React from "react";
 import styled from "styled-components";
 
-import plus from "../UI/counter-plus.svg";
-import minus from "../UI/counter-minus.svg";
+import plus from "../../UI/counter-plus.svg";
+import minus from "../../UI/counter-minus.svg";
 
 const Guests = styled.div`
   padding: 10px 14px;
