+++ src/Homes/All/Filters/PopUp.js
@@ -2,7 +2,7 @@ import React from "react";
 import styled from "styled-components";
 import MediaQuery from "react-responsive";
 
-import close from "../UI/close.svg";
+import close from "../../UI/close.svg";
 
 const Wrapper = styled.div`
   width: 100%;
