+++ src/Homes/Room/index.js
@@ -7,9 +7,9 @@ import styled from "styled-components";
 import Nav from "./Nav";
 import Gallery from "./Gallery";
 import BookRequest from "./BookRequest";
-import Overview from "./Overview";
+import Overview from "./Overview/";
 import Reviews from "./Reviews/";
-import Host from "./Host";
+import Host from "./Host/";
 import Location from "./Location";
 import Offers from "./Offers/";
 import Explore from "./Explore";
